// src/generated/prisma/models/BlogPost.ts
export type BlogPostWhereInput = {
  AND?: Prisma.BlogPostWhereInput | Prisma.BlogPostWhereInput[]
  OR?: Prisma.BlogPostWhereInput[]
  NOT?: Prisma.BlogPostWhereInput | Prisma.BlogPostWhereInput[]
  id?: Prisma.StringFilter<"BlogPost"> | string
  type?: Prisma.EnumBlogPostTypeFilter<"BlogPost"> | $Enums.BlogPostType
  createdAt?: Prisma.DateTimeFilter<"BlogPost"> | Date | string
  updatedAt?: Prisma.DateTimeFilter<"BlogPost"> | Date | string
  title?: Prisma.StringFilter<"BlogPost"> | string
  slug?: Prisma.StringFilter<"BlogPost"> | string
  content?: Prisma.StringFilter<"BlogPost"> | string
  excerpt?: Prisma.StringFilter<"BlogPost"> | string
  imageUrl?: Prisma.StringNullableFilter<"BlogPost"> | string | null
  author?: Prisma.StringFilter<"BlogPost"> | string
  published?: Prisma.BoolFilter<"BlogPost"> | boolean
  publishedAt?: Prisma.DateTimeNullableFilter<"BlogPost"> | Date | string | null
  locale?: Prisma.StringFilter<"BlogPost"> | string
  tags?: Prisma.StringFilter<"BlogPost"> | string
  metaTitle?: Prisma.StringFilter<"BlogPost"> | string
  metaDesc?: Prisma.StringFilter<"BlogPost"> | string
}